@@ -16,7 +16,7 @@
 
 <script>
 export default {
-  name: "ProductsListItem",
+  name: "ProductsListItemComponent",
   props: ["product"],
   data() {
     return {
